@@ -4,15 +4,15 @@ import { Link } from 'react-router';
 import agent from '../agent';
 import { connect } from 'react-redux';
 import {
-  PROFILE_FAVORITES_PAGE_LOADED,
-  PROFILE_FAVORITES_PAGE_UNLOADED
+  PROFILE_PAGE_LOADED,
+  PROFILE_PAGE_UNLOADED
 } from '../constants/actionTypes';
 
 const mapDispatchToProps = dispatch => ({
   onLoad: (pager, payload) =>
-    dispatch({ type: PROFILE_FAVORITES_PAGE_LOADED, pager, payload }),
+    dispatch({ type: PROFILE_PAGE_LOADED, pager, payload }),
   onUnload: () =>
-    dispatch({ type: PROFILE_FAVORITES_PAGE_UNLOADED })
+    dispatch({ type: PROFILE_PAGE_UNLOADED })
 });
 
 class ProfileFavorites extends Profile {
